requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
rank-bm25
sentence-transformers
torch
qdrant-client
python-dotenv
redis
openai

